Moved client total measurement to own component
