+++ Server/Api.Websocket/Handlers/AirQualityGraphEventHandler.cs
@@ -3,11 +3,13 @@ using WebSocketBoilerplate;
 
 namespace Api.Websocket.Handlers;
 
+// hvad klient sender
 public class RequestAirQualityDataDto : BaseDto
 {
     public string SomethingTheClientSends { get; set; }
 }
 
+// hvad server sender..
 public class AirQualityDataGraph : BaseDto
 {
     public string SomethingTheServerSends { get; set; }
